Add documentation and new method to FrontendScene
I decided to make it so that each FrontendScene has a show() method that
causes that scene to be shown. Since each scene should have access to
its parent Frontend (which is expected to store things like the Stage in
a JavaFX-based Frontend), this should be easy to implement.

This feels limiting but I cannot yet see how, so to avoid excessive
pondering I've decided to just go with it.